Bugfix to evert()
It's supposed to be a nice side-effect of evert() that not only is
the target vertex the real root after this operaiton, it's also the
root of its own path tree.

Calling reverse() on its right child spoils this and also causes
problems with the remainder of the implementation because the first
thing that routine does is splay the node-to-be-reversed.